cloud_computing/batikrecog-86-cloud/requirements.txt
firebase-admin==5.0.0
Flask
imageio
jsonpickle
numpy
opencv-python-headless
Pillow
tensorflow-cpu==2.5.0